add double linked list data structure in Swift
